refactor(Header): extracting Save button and adding all new buttons to the Button component exports
