+++ src/components/Button/Add/ButtonAdd.styles.ts
@@ -1,7 +1,7 @@
 import { StyleSheet, ViewStyle } from 'react-native'
 
-import { Colors, colors } from '../../styles'
-import * as StyleUtils from '../../styles/styles.utils'
+import { Colors, colors } from '../../../styles'
+import * as StyleUtils from '../../../styles/styles.utils'
 
 const styles = {
   button: {
