+++ src/components/Button/Save/ButtonSave.styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet, ViewStyle } from 'react-native'
+
+const styles = {
+  button: {
+    flexDirection: 'row',
+    paddingRight: 20,
+  } as ViewStyle,
+}
+
+export default StyleSheet.create(styles)
