+++ src/components/Button/Save/index.ts
@@ -1,0 +1,3 @@
+import ButtonSave from './ButtonSave'
+
+export default ButtonSave
