+++ src/components/Button/index.ts
@@ -1,5 +1,8 @@
+import ButtonAdd from './Add'
+import ButtonBack from './Back'
 import Button from './Button'
 import { ButtonSizes, ButtonVariants } from './Button.types'
+import ButtonSave from './Save'
 
 export default Button
-export { ButtonVariants, ButtonSizes }
+export { ButtonAdd, ButtonBack, ButtonSave, ButtonVariants, ButtonSizes }
