+++ src/components/Header/Header.tsx
@@ -1,13 +1,10 @@
 import { StackNavigationProp } from '@react-navigation/stack'
-import Button, { ButtonVariants } from 'components/Button'
 import React from 'react'
-import { useTranslation } from 'react-i18next'
 import { View } from 'react-native'
-import { Colors } from 'styles'
 
 import { types as HomeNavigationTypes } from '../../modules/HomeNavigation'
-import ButtonAdd from '../ButtonAdd'
-import ButtonBack from '../ButtonBack'
+import { Colors } from '../../styles'
+import { ButtonAdd, ButtonBack, ButtonSave } from '../Button'
 import Optional from '../Optional'
 import Text, { HeaderLevels } from '../Typography'
 import styles from './Header.styles'
@@ -30,41 +27,27 @@ const Header = ({
   isSaveButtonEnabled = false,
   onAdd,
   actionItem,
-}: Props) => {
-  const { t } = useTranslation()
-
-  return (
-    <View style={styles.container}>
-      <ButtonBack onPress={navigation.goBack} />
-      <Text.Header level={HeaderLevels.H5} color={Colors.PrimaryPurple}>
-        {headerText}
-      </Text.Header>
-      <Optional
-        // TODO: refactoring to only expect the actionItem
-        condition={!!actionItem}
-        fallback={
-          <Optional
-            condition={showAddButton}
-            fallback={
-              <Button
-                isDisabled={!isSaveButtonEnabled}
-                variant={ButtonVariants.Clear}
-                label={t('Save')}
-                color={Colors.PrimaryGreen}
-                onPress={onSave}
-                style={styles.button}
-                isFullWidth={false}
-              />
-            }
-          >
-            <ButtonAdd onPress={onAdd!} />
-          </Optional>
-        }
-      >
-        {actionItem}
-      </Optional>
-    </View>
-  )
-}
+}: Props) => (
+  <View style={styles.container}>
+    <ButtonBack onPress={navigation.goBack} />
+    <Text.Header level={HeaderLevels.H5} color={Colors.PrimaryPurple}>
+      {headerText}
+    </Text.Header>
+    <Optional
+      // TODO: refactoring to only expect the actionItem
+      condition={!!actionItem}
+      fallback={
+        <Optional
+          condition={showAddButton}
+          fallback={<ButtonSave onPress={onSave!} isDisabled={!isSaveButtonEnabled} />}
+        >
+          <ButtonAdd onPress={onAdd!} />
+        </Optional>
+      }
+    >
+      {actionItem}
+    </Optional>
+  </View>
+)
 
 export default Header
